% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{clbp}
\alias{clbp}
\title{Rubinstein et al. (2019)'s chronic low back pain data}
\format{A data frame with 23 rows and 8 variables}
\usage{
data(clbp)
}
\description{
\itemize{
  \item \code{ID}: Study ID
  \item \code{Souce}: First author name and year of publication
  \item \code{m1}: Estimated mean in experimental group
  \item \code{s1}: Standard deviation in experimental group
  \item \code{n1}: Number of observations in experimental group
  \item \code{m2}: Estimated mean in control group
  \item \code{s2}: Standard deviation in control group
  \item \code{n2}: Number of observations in control group
}
}
\references{
Rubinstein, S. M,, de Zoete, A., van Middelkoop, M.,
Assendelft, W. J. J., de Boer, M. R., van Tulder, M. W. (2019).
Benefits and harms of spinal manipulative therapy for
the treatment of chronic low back pain: systematic review
and meta-analysis of randomised controlled trials.
\emph{BMJ.}
\strong{364}: l689.
\url{https://doi.org/10.1136/bmj.l689}
}
\keyword{datasets}
