% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{hyp}
\alias{hyp}
\title{Hypertension data}
\format{A data frame with 10 rows and 2 variables}
\usage{
data(hyp)
}
\description{
The hypertension data (Wang et al., 2005) included
7 studies comparing the treatment effect of
anti-hypertensive treatment versus control on reducing
diastolic blood pressure (DBP) in patients with hypertension.
Negative estimates indicate the reduction of DBP in the
anti-hypertensive treatment group.
}
\details{
\itemize{
  \item \code{y}: Standardized mean difference
  \item \code{se}: Standard error
  \item \code{label}: Labels for each study
}
}
\references{
Wang, J. G., Staessen, J. A., Franklin, S. S., Fagard, R.,
and Gueyffier, F. (2005).
Systolic and diastolic blood pressure lowering as determinants
of cardiovascular outcome.
\emph{Hypertension.}
\strong{45}(5): 907-913.
\url{https://doi.org/10.1161/01.HYP.0000165020.14745.79}
}
\keyword{datasets}
