% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-mu_f.r
\name{mu_f}
\alias{mu_f}
\title{Generate posterior distribution of \eqn{mu}}
\usage{
mu_f(fit, plot = TRUE)
}
\arguments{
\item{fit}{An object of S4 class \code{stanfit} representing
the fitted results.}

\item{plot}{logical; if \code{TRUE}, then this function plots
posterior distribution of \eqn{mu}.}
}
\description{
Generate posterior distribution of \eqn{mu}
}
\examples{
\donttest{
require("flexmeta")
require("rstan")
options(mc.cores = parallel::detectCores())

data("clbp")
dat <- convert_mean(clbp$n1, clbp$m1, clbp$s1, clbp$n2, clbp$m2, clbp$s2)
dat <- list(K = length(dat$y), y = dat$y, se = dat$se)

fit <- sampling(sm_sn, data = dat,
  chains = 1, warmup = 5000, iter = 30000, seed = 2645751)

mu_f(fit)
}
}
\seealso{
\code{\link[rstan]{sampling}}, \code{\link[rstan]{stanmodel-class}},
  \code{\link[rstan]{stanfit-class}}, \code{\link[rstan]{stan}}.
}
