% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nfld}
\alias{nfld}
\title{Koutoukidis et al. (2019)'s nonalcoholic fatty liver disease data}
\format{A data frame with 25 rows and 8 variables}
\usage{
data(nfld)
}
\description{
\itemize{
  \item \code{ID}: Study ID
  \item \code{Souce}: First author name and year of publication
  \item \code{m1}: Estimated mean in experimental group
  \item \code{s1}: Standard deviation in experimental group
  \item \code{n1}: Number of observations in experimental group
  \item \code{m2}: Estimated mean in control group
  \item \code{s2}: Standard deviation in control group
  \item \code{n2}: Number of observations in control group
}
}
\references{
Koutoukidis, D. A,, Astbury, N. M., Tudor, K. E.,
Morris, E., Henry, J. A., Noreik, M., Jebb, S. A.,
Aveyard, P. (2019).
Association of Weight Loss Interventions With Changes
in Biomarkers of Nonalcoholic Fatty Liver Disease:
A Systematic Review and Meta-analysis.
\emph{JAMA Intern Med.}
\strong{179}(9): 1262-1271.
\url{https://doi.org/10.1001/jamainternmed.2019.2248}
}
\keyword{datasets}
