% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pain}
\alias{pain}
\title{Pain data}
\format{A data frame with 22 rows and 2 variables}
\usage{
data(pain)
}
\description{
The pain data (Riley et al., 2011; Hauser et al., 2009) included
22 studies comparing the treatment effect of antidepressants on
reducing pain in patients with fibromyalgia syndrome.
The treatment effects were summarized using standardized mean
differences on a visual analog scale for pain between the
antidepressant group and control group.
Negative estimates indicate the reduction of pain in the
antidepressant group.
}
\details{
\itemize{
  \item \code{y}: Standardized mean difference
  \item \code{sigmak}: Standard error
}
}
\references{
Hauser, W., Bernardy, K, Uceyler, N., and Sommer, C. (2009).
Treatment of fibromyalgia syndrome with antidepressants: a meta-analysis.
\emph{JAMA.}
\strong{301}(2): 198-209.
\url{https://jamanetwork.com/journals/jama/fullarticle/183189}

Riley, R. D., Higgins, J. P. T, and Deeks, J. J. (2011).
Interpretation of random-effects meta-analyses.
\emph{BMJ.}
\strong{342}: d549.
\url{https://doi.org/10.1136/bmj.d549}
}
\keyword{datasets}
