% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-pred_f.r
\name{pred_f}
\alias{pred_f}
\title{Generate Summary of Predictive Distribution}
\usage{
pred_f(fit, plot = TRUE)
}
\arguments{
\item{fit}{An object of S4 class \code{stanfit} representing
the fitted results.}

\item{plot}{logical; if \code{TRUE}, then this function plots
predictive distribution of \eqn{\theta_{new}}.}
}
\description{
Generate Summary of Predictive Distribution
}
\examples{
require("flexmeta")
}
