% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-stan.R
\name{sas_stan}
\alias{sas_stan}
\title{Bayesian meta-analysis using a Sinh-arcsinh prior}
\usage{
sas_stan(data, par = NA, chains = 1, iter = 3500, warmup = 500,
  thin = 1, init = "random")
}
\arguments{
\item{data}{A list of the data}

\item{par}{parameters of interest}

\item{chains}{the number of Markov chains}

\item{iter}{the number of iterations for each chain including warmup}

\item{warmup}{the number of warmup iterations per chain}

\item{thin}{the period for saving samples}

\item{init}{A list of the initial values}
}
\value{
\itemize{
\item \code{f}: An object of S4 class \code{stanfit} representing the fitted results.
Slot \code{mode} for this object indicates if the sampling is done or not.
}
}
\description{
Bayesian meta-analysis using a Sinh-arcsinh prior
}
\examples{
require("rstansas")
m1 <- c(15,12,29,42,14,44,14,29,10,17,38,19,21)
n1 <- c(16,16,34,56,22,54,17,58,14,26,44,29,38)
m2 <- c( 9, 1,18,31, 6,17, 7,23, 3, 6,12,22,19)
n2 <- c(16,16,34,56,22,55,15,58,15,27,45,30,38)
dat1 <- pimeta::convert_bin(m1, n1, m2, n2, type = "logOR")
dat2 <- list(K = dim(dat1)[1], yi = dat1$y, si = dat1$se)
init1 <- function() {
  list(sigma = runif(1, 0, 2), delta = runif(1, 0, 2),
       mu = runif(1, -2, 2), epsilon = runif(1, -2, 2),
       theta = runif(dim(dat1)[1], -2, 2), theta_new = runif(1, -2, 2))
}
sas_stan(data = dat2, init = init1)
}
