% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sbp}
\alias{sbp}
\title{Systolic blood pressure (SBP) data}
\format{A data frame with 10 rows and 2 variables}
\usage{
data(sbp)
}
\description{
Riley et al. (2011) analyzed a hypothetical meta-analysis.
They generated a data set of 10 studies examining the same
antihypertensive drug.
Negative estimates suggested reduced blood pressure in the
treatment group.
}
\details{
\itemize{
  \item \code{y}: Standardized mean difference
  \item \code{sigmak}: Standard error
  \item \code{label}: Labels for each generated study
}
}
\references{
Riley, R. D., Higgins, J. P. T, and Deeks, J. J. (2011).
Interpretation of random-effects meta-analyses.
\emph{BMJ.}
\strong{342}: d549.
\url{https://doi.org/10.1136/bmj.d549}
}
\keyword{datasets}
