% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{setshift}
\alias{setshift}
\title{Set-shifting data}
\format{A data frame with 14 rows and 2 variables}
\usage{
data(setshift)
}
\description{
Higgins et al. (2009) re-analyzed data (Roberts et al., 2007)
that included 14 studies evaluating the set-shifting ability
in people with eating disorders by using a prediction interval.
Standardized mean differences in the time taken to complete
Trail Making Test between subjects with eating disorders and
healthy controls were collected.
Positive estimates indicate impairment in set shifting ability
in people with eating disorders.
}
\details{
\itemize{
  \item \code{y}: Standardized mean difference
  \item \code{sigmak}: Standard error
}
}
\references{
Roberts, M. E., Tchanturia, K., Stahl, D.,
Southgate, L., and Treasure, J. (2007).
A systematic review and meta-analysis of set-shifting
ability in eating disorders.
\emph{Psychol Med.}
\strong{37}(8): 1075-1084.
\url{https://doi.org/10.1017/S0033291707009877}

Higgins, J. P. T, Thompson, S. G., Spiegelhalter, D. J. (2009).
A re-evaluation of random-effects meta-analysis.
\emph{J R Stat Soc Ser A Stat Soc.}
\strong{172}(1): 137-159.
\url{https://doi.org/10.1111/j.1467-985X.2008.00552.x}
}
\keyword{datasets}
