% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-V_f.r
\name{V_f}
\alias{V_f}
\title{Generate posterior distribution of \eqn{V}}
\usage{
V_f(fit, plot = TRUE)
}
\arguments{
\item{fit}{An object of S4 class \code{stanfit} representing
the fitted results.}

\item{plot}{logical; if \code{TRUE}, then this function plots
posterior distribution of \eqn{V}.}
}
\description{
Generate posterior distribution of \eqn{V}
}
\examples{
require("flexmeta")
}
