% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-dic.r
\name{dic}
\alias{dic}
\title{Generate DIC}
\usage{
dic(fit, dat)
}
\arguments{
\item{fit}{An object of S4 class \code{stanfit} representing
the fitted results.}

\item{dat}{A list of the study data.
\itemize{
\item \code{y}: A vector of the effect size estimates.
\item \code{se}: A vector of the within studies standard error
  estimates.
}}
}
\value{
A numeric value of \eqn{DIC}.
}
\description{
Generate DIC
}
\examples{
require("flexmeta")
}
\seealso{
stanmodels
}
