% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-mu_f.r
\name{mu_f}
\alias{mu_f}
\title{Generate posterior distribution of \eqn{mu}}
\usage{
mu_f(fit, plot = TRUE)
}
\arguments{
\item{fit}{An object of S4 class \code{stanfit} representing
the fitted results.}

\item{plot}{logical; if \code{TRUE}, then this function plots
posterior distribution of \eqn{mu}.}
}
\description{
Generate posterior distribution of \eqn{mu}
}
\examples{
require("flexmeta")
}
